requirements.txt
Django==1.8
django-cors-headers==1.0.0
django-filter==0.9.2
djangorestframework==3.1.1
Markdown==2.6.1
pytz==2015.2
